21_sqlalchemy_project_exr/requirements.txt
alembic==1.13.2
greenlet==3.0.3
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
python-decouple==3.8
SQLAlchemy==2.0.31
typing_extensions==4.12.2
